move files into sub directory
