+++ vstgui/lib/viewlayouter/autosizeviewlayouter.cpp
