+++ vstgui/lib/viewlayouter/autosizeviewlayouter.h
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "iviewlayouter.h"
+#include "../iviewlayouter.h"
 #include <tuple>
 #include <vector>
